remove unrelated changes, keep PR focused on unknown module type
