+++ packages/node-modules-inspector/src/node/cli.ts
@@ -113,33 +113,5 @@ cli
     })
   })
 
-cli
-  .command('check', 'Run analysis and config hook without starting the server (for CI/CD)')
-  .option('--root <dir>', 'Root directory', { default: process.cwd() })
-  .option('--config <file>', 'Config file')
-  .option('--depth <depth>', 'Max depth to list dependencies', { default: 8 })
-  .action(async (options) => {
-    const { createServerFunctions } = await import('./rpc')
-    const { storageNpmMeta, storageNpmMetaLatest, storagePublint } = await import('./storage')
-
-    const fns = createServerFunctions({
-      cwd: options.root,
-      depth: Number(options.depth),
-      configFile: options.config,
-      mode: 'build',
-      storageNpmMeta,
-      storageNpmMetaLatest,
-      storagePublint,
-    })
-
-    try {
-      await fns.getPayload()
-    }
-    catch (error: any) {
-      console.error(c.red`✖ ${error.message || error}`)
-      process.exit(1)
-    }
-  })
-
 cli.help()
 cli.parse()
