--- packages/node-modules-tools/src/agents/rush/index.ts
@@ -1,1 +1,0 @@
-export { listPackageDependencies } from './list'
